@@ -35,10 +35,12 @@ setup(
         "all": (TEST_REQS, EXTRA_REQS),
     },
     tests_require=(TEST_REQS),
-    python_requires=">=3.6",
+    python_requires=">=3.8",
     classifiers=(
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ),
     keywords="groundwater data",
 )
